requirements.txt
absl-py==0.11.0
astor==0.8.1
cached-property==1.5.2
cycler==0.10.0
gast==0.2.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==3.1.0
importlib-metadata==3.1.0
Keras==2.1.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
numpy==1.19.3
opt-einsum==3.3.0
Pillow==8.0.1
protobuf==3.14.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
scipy==1.5.4
six==1.15.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
